# Return UNESCAPED_CHARS immediately