Doing this gives a very small (but real) performance improvement in stage 1 on Kady Lake (Haswell+), and also reduces the amount of indirection, which will improve readability/grokkability.

Note: I attempted to do the same thing with the utf8 check, but that had a negative effect. Presumably this is because the UTF-8 error mask is in SIMD, and it's a lot more expensive to pull data from SIMD to scalar world with operations like test and movmskb. unescaped_chars_error was already scalar and didn't have that problem.