@jkeiser 2019-10-18
Also interesting: this is one of the cases where `unlikely` was absolutely necessary. I think the rule is probably *any* work in either the if condition or the conditional block has impact unless you use `unlikely`. Passthrough if statements (like `if (error) { return error; }`) do not require `likely`.