@lemire 2019-10-21
@jkeiser 

I think we need to be prudent when it comes to replacing branchless code with branchy code because benchmarks can easily favor branchy code in benchmarks for misleading reasons.  See my post [Benchmarking is hard: processors learn to predict branches](https://lemire.me/blog/2019/10/16/benchmarking-is-hard-processors-learn-to-predict-branches/).

Do we have a rationale for why we think that your code should be more performant?


In at least one instance, I see a performance degradation. On a skylake processor with GCC 8.3.

Before:
```
jsonexamples/twitter.json
stage 1 instructions:    1711881 cycles:     481401 (48.86 %) ins/cycles: 3.56 mis. branches:        982 (cycles/mis.branch 490.06) cache accesses:      29169 (failure          0)
 stage 1 runs at 0.76 cycles per input byte.
```



After:
```
jsonexamples/twitter.json
stage 1 instructions:    1728989 cycles:     489486 (49.02 %) ins/cycles: 3.53 mis. branches:       1234 (cycles/mis.branch 396.46) cache accesses:      29230 (failure          0)
 stage 1 runs at 0.78 cycles per input byte.
```



Here is the full output:


Before:

Trial 1
```
number of iterations 1000 
number of bytes 631515 number of structural chars 55264 ratio 0.088
mem alloc instructions:       1197 cycles:        740 (0.08 %) ins/cycles: 1.62 mis. branches:          3 (cycles/mis.branch 239.76) cache accesses:      29160 (failure          0)
 mem alloc runs at 0.00 cycles per input byte.
stage 1 instructions:    1711881 cycles:     480798 (48.98 %) ins/cycles: 3.56 mis. branches:        989 (cycles/mis.branch 486.07) cache accesses:      29160 (failure          2)
 stage 1 runs at 0.76 cycles per input byte.
stage 2 instructions:    1490176 cycles:     500122 (50.95 %) ins/cycles: 2.98 mis. branches:       1522  (cycles/mis.branch 328.48)  cache accesses:      54494 (failure          5)
 stage 2 runs at 0.79 cycles per input byte and 9.05 cycles per structural character.
 all stages: 1.55 cycles per input byte.
Estimated average frequency: 3.715 GHz.
Min:  0.000264275 bytes read: 631515 Gigabytes/second: 2.38961
```

Trial 2
```
number of iterations 1000 
number of bytes 631515 number of structural chars 55264 ratio 0.088
mem alloc instructions:       1197 cycles:        725 (0.07 %) ins/cycles: 1.65 mis. branches:          2 (cycles/mis.branch 289.09) cache accesses:      29112 (failure          0)
 mem alloc runs at 0.00 cycles per input byte.
stage 1 instructions:    1711881 cycles:     481567 (48.88 %) ins/cycles: 3.55 mis. branches:        992 (cycles/mis.branch 485.11) cache accesses:      29112 (failure          0)
 stage 1 runs at 0.76 cycles per input byte.
stage 2 instructions:    1490176 cycles:     502892 (51.05 %) ins/cycles: 2.96 mis. branches:       1529  (cycles/mis.branch 328.74)  cache accesses:      54562 (failure          0)
 stage 2 runs at 0.80 cycles per input byte and 9.10 cycles per structural character.
 all stages: 1.56 cycles per input byte.
Estimated average frequency: 3.714 GHz.
Min:  0.000265279 bytes read: 631515 Gigabytes/second: 2.38057
```


After:

Trial 1
```
number of iterations 1000 
number of bytes 631515 number of structural chars 55264 ratio 0.088
mem alloc instructions:       1197 cycles:        840 (0.08 %) ins/cycles: 1.42 mis. branches:          4 (cycles/mis.branch 209.49) cache accesses:      29254 (failure          0)
 mem alloc runs at 0.00 cycles per input byte.
stage 1 instructions:    1728989 cycles:     493457 (49.19 %) ins/cycles: 3.50 mis. branches:       1231 (cycles/mis.branch 400.84) cache accesses:      29254 (failure          0)
 stage 1 runs at 0.78 cycles per input byte.
stage 2 instructions:    1490176 cycles:     508791 (50.72 %) ins/cycles: 2.93 mis. branches:       1780  (cycles/mis.branch 285.82)  cache accesses:      54636 (failure          0)
 stage 2 runs at 0.81 cycles per input byte and 9.21 cycles per structural character.
 all stages: 1.59 cycles per input byte.
Estimated average frequency: 3.728 GHz.
Min:  0.000269056 bytes read: 631515 Gigabytes/second: 2.34715
```

Trial 2
```
number of iterations 1000 
number of bytes 631515 number of structural chars 55264 ratio 0.088
mem alloc instructions:       1197 cycles:        778 (0.08 %) ins/cycles: 1.54 mis. branches:          2 (cycles/mis.branch 312.16) cache accesses:      29231 (failure          0)
 mem alloc runs at 0.00 cycles per input byte.
stage 1 instructions:    1728989 cycles:     489177 (48.91 %) ins/cycles: 3.53 mis. branches:       1232 (cycles/mis.branch 396.84) cache accesses:      29231 (failure          0)
 stage 1 runs at 0.77 cycles per input byte.
stage 2 instructions:    1490176 cycles:     510193 (51.01 %) ins/cycles: 2.92 mis. branches:       1780  (cycles/mis.branch 286.59)  cache accesses:      54683 (failure          1)
 stage 2 runs at 0.81 cycles per input byte and 9.23 cycles per structural character.
 all stages: 1.58 cycles per input byte.
Estimated average frequency: 3.730 GHz.
Min:  0.000268172 bytes read: 631515 Gigabytes/second: 2.35489
```


So I suggest we go forward by examining the benchmarks more closely. What are your numbers telling you?