@jkeiser 2019-10-21
I'll send the bench info over when I get home later.

I tried this because I noticed an improvement when I changed an error if statement to. My reasoning is that branch prediction of branches we never take (in the normal course) ought to be near enough to perfect as to be noise. As far as I understand, it's not branches that are the problem, it's mispredicts, right?

I'm ignorant of how compilers optimize `unlikely`, but my expectation is it lets the compiler know it can safely move the if block out of the code cache line so more likely instructions can be decoded earlier, and probably lets it schedule the if condition later than usual to allow more data critical statements to execute earlier.

Using a return means that we break a data dependency (the error variable) between one iteration of the loop and the next, allowing them to run somewhat more out of order, too.

I was surprised when I discovered this, but I think it makes sense. It'll be interesting to look what's going on in Sandy bridge!

I'll mark it engineering research as well, since it sounds like it's not a clear improvement across all architectures. It's really not a large enough improvement that we'd want to drop everything to get it in, anyway :)