@lemire 2019-11-01
> Using a return instead of error |= more_errors also means that we break a data dependency between one iteration of the loop and the next, allowing them to run somewhat more out of order, too.

You may be right, but I am skeptical. Can you elaborate?