@lemire 2019-11-01
> I'm ignorant of how compilers optimize unlikely, but my layman's expectation is unlikely lets the compiler know it can safely move the if block out of the code cache line so more likely instructions can be decoded earlier. It also probably lets it schedule the if condition instructions later than usual to allow more critical statements to execute earlier.

I think that what it does, often, is just to move out the instructions corresponding to the unlikely choice entirely off the main code sequence so that no jump is typically needed and you have a very compact code sequence (without the unlikely instructions occupying cache memory or requiring decoding). Of course, who knows what else it might do...?